data-engineering/pyproject.toml
[tool.poetry]
name = "data-engineering"
version = "0.1.0"
description = ""
authors = ["Rithwik Babu <rbabu4@wisc.edu>"]
readme = "README.md"

[[tool.poetry.packages]]
include = "pipeline"
from = "src"

[[tool.poetry.packages]]
include = "api"
from = "src"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "2.2.3"
pydantic = "^2.10.6"
mypy = "^1.14.1"
pyyaml = "^6.0.2"
fds-sdk-formula = "^3.0.1"
fds-sdk-utils = "^2.0.1"
google-cloud-bigquery = "^3.29.0"
google-cloud-bigquery-storage = "^2.27.0"
db-dtypes = "^1.4.0"
google-cloud-storage = "^3.0.0"
pandas-market-calendars = "^4.4.0"
modal = "^0.73.116"
pymsteams = "^0.2.5"
pandas-gbq = "^0.28.0"
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
opentelemetry-api = "^1.35.0"
opentelemetry-sdk = "^1.35.0"
opentelemetry-exporter-otlp = "^1.35.0"
hawk-sdk = "^0.0.20"
fsspec = "^2025.7.0"
gcsfs = "^2025.7.0"


[tool.poetry.group.dev.dependencies]
mkdocs = "^1.6.1"
mkdocs-material = "^9.6.2"
mkdocstrings = "^0.28.0"
mkdocstrings-python = "^1.14.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
